fix refs #2632: Keep Lark notification delivery out of authoring
Move credential-bearing Lark notification ports to scheduled composition and keep card rendering in the platform package so Authoring.Lark remains authoring-only.
